requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dateparser==1.1.8
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-mail==1.2.8
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
passlib==1.7.4
psycopg2-binary==2.9.6
PuLP==2.7.0
pyasn1==0.5.0
pycryptodome==3.18.0
pydantic==1.10.8
python-binance==1.0.17
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
regex==2023.5.5
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
